update script generation names
