+++ src/holon/serializers/datamodel/mapper.py
@@ -1,4 +1,3 @@
-
 ###################################################
 ## Note! This script is automatically generated! ##
 ###################################################
@@ -6,133 +5,120 @@
 from rest_polymorphic.serializers import PolymorphicSerializer
 
 
-from holon.models.actor import (
-        NonFirmActor
-)
+from holon.models.actor import NonFirmActor
 
 
 from holon.models.contract import (
-        DeliveryContract,
-        ConnectionContract,
-        TaxContract,
-        TransportContract
+    DeliveryContract,
+    ConnectionContract,
+    TaxContract,
+    TransportContract,
 )
 
 
 from holon.models.gridconnection import (
-        BuiltEnvironmentGridConnection,
-        UtilityGridConnection,
-        HouseGridConnection,
-        BuildingGridConnection,
-        ProductionGridConnection,
-        IndustryGridConnection,
-        DistrictHeatingGridConnection
+    BuiltEnvironmentGridConnection,
+    UtilityGridConnection,
+    HouseGridConnection,
+    BuildingGridConnection,
+    ProductionGridConnection,
+    IndustryGridConnection,
+    DistrictHeatingGridConnection,
 )
 
 
 from holon.models.asset import (
-        ConsumptionAsset,
-        DieselVehicleAsset,
-        HeatConsumptionAsset,
-        ElectricConsumptionAsset,
-        HybridConsumptionAsset,
-        ConversionAsset,
-        VehicleConversionAsset,
-        ElectricCoversionAsset,
-        CookingConversionAsset,
-        HeatConversionAsset,
-        ChemicalHeatConversionAsset,
-        ElectricHeatConversionAsset,
-        TransportHeatConversionAsset,
-        HybridHeatCoversionAsset,
-        ProductionAsset,
-        ElectricProductionAsset,
-        HeatProductionAsset,
-        HybridProductionAsset,
-        StorageAsset,
-        HeatStorageAsset,
-        ElectricStorageAsset,
-        VehicleElectricStorageAsset
+    ConsumptionAsset,
+    DieselVehicleAsset,
+    HeatConsumptionAsset,
+    ElectricConsumptionAsset,
+    HybridConsumptionAsset,
+    ConversionAsset,
+    VehicleConversionAsset,
+    ElectricCoversionAsset,
+    CookingConversionAsset,
+    HeatConversionAsset,
+    ChemicalHeatConversionAsset,
+    ElectricHeatConversionAsset,
+    TransportHeatConversionAsset,
+    HybridHeatCoversionAsset,
+    ProductionAsset,
+    ElectricProductionAsset,
+    HeatProductionAsset,
+    HybridProductionAsset,
+    StorageAsset,
+    HeatStorageAsset,
+    ElectricStorageAsset,
+    VehicleElectricStorageAsset,
 )
 
 
-from holon.models.gridnode import (
-        ElectricGridNode,
-        HeatGridNode
-)
+from holon.models.gridnode import ElectricGridNode, HeatGridNode
 
 
+from .subserializers import NonFirmActorSerializer
 
 
 from .subserializers import (
-        NonFirmActorSerializer
-)
-
-
-from .subserializers import (
-        DeliveryContractSerializer,
-        ConnectionContractSerializer,
-        TaxContractSerializer,
-        TransportContractSerializer
+    DeliveryContractSerializer,
+    ConnectionContractSerializer,
+    TaxContractSerializer,
+    TransportContractSerializer,
 )
 
 
 from .subserializers import (
-        BuiltEnvironmentGridConnectionSerializer,
-        UtilityGridConnectionSerializer,
-        HouseGridConnectionSerializer,
-        BuildingGridConnectionSerializer,
-        ProductionGridConnectionSerializer,
-        IndustryGridConnectionSerializer,
-        DistrictHeatingGridConnectionSerializer
+    BuiltEnvironmentGridConnectionSerializer,
+    UtilityGridConnectionSerializer,
+    HouseGridConnectionSerializer,
+    BuildingGridConnectionSerializer,
+    ProductionGridConnectionSerializer,
+    IndustryGridConnectionSerializer,
+    DistrictHeatingGridConnectionSerializer,
 )
 
 
 from .subserializers import (
-        ConsumptionAssetSerializer,
-        DieselVehicleAssetSerializer,
-        HeatConsumptionAssetSerializer,
-        ElectricConsumptionAssetSerializer,
-        HybridConsumptionAssetSerializer,
-        ConversionAssetSerializer,
-        VehicleConversionAssetSerializer,
-        ElectricCoversionAssetSerializer,
-        CookingConversionAssetSerializer,
-        HeatConversionAssetSerializer,
-        ChemicalHeatConversionAssetSerializer,
-        ElectricHeatConversionAssetSerializer,
-        TransportHeatConversionAssetSerializer,
-        HybridHeatCoversionAssetSerializer,
-        ProductionAssetSerializer,
-        ElectricProductionAssetSerializer,
-        HeatProductionAssetSerializer,
-        HybridProductionAssetSerializer,
-        StorageAssetSerializer,
-        HeatStorageAssetSerializer,
-        ElectricStorageAssetSerializer,
-        VehicleElectricStorageAssetSerializer
+    ConsumptionAssetSerializer,
+    DieselVehicleAssetSerializer,
+    HeatConsumptionAssetSerializer,
+    ElectricConsumptionAssetSerializer,
+    HybridConsumptionAssetSerializer,
+    ConversionAssetSerializer,
+    VehicleConversionAssetSerializer,
+    ElectricCoversionAssetSerializer,
+    CookingConversionAssetSerializer,
+    HeatConversionAssetSerializer,
+    ChemicalHeatConversionAssetSerializer,
+    ElectricHeatConversionAssetSerializer,
+    TransportHeatConversionAssetSerializer,
+    HybridHeatCoversionAssetSerializer,
+    ProductionAssetSerializer,
+    ElectricProductionAssetSerializer,
+    HeatProductionAssetSerializer,
+    HybridProductionAssetSerializer,
+    StorageAssetSerializer,
+    HeatStorageAssetSerializer,
+    ElectricStorageAssetSerializer,
+    VehicleElectricStorageAssetSerializer,
 )
 
 
-from .subserializers import (
-        ElectricGridNodeSerializer,
-        HeatGridNodeSerializer
-)
-
+from .subserializers import ElectricGridNodeSerializer, HeatGridNodeSerializer
 
 
 class ActorPolymorphicSerializer(PolymorphicSerializer):
-    model_serializer_mapping = {
-        NonFirmActor: NonFirmActorSerializer
-    }
+    model_serializer_mapping = {NonFirmActor: NonFirmActorSerializer}
+
 
 class ContractPolymorphicSerializer(PolymorphicSerializer):
     model_serializer_mapping = {
         DeliveryContract: DeliveryContractSerializer,
         ConnectionContract: ConnectionContractSerializer,
         TaxContract: TaxContractSerializer,
-        TransportContract: TransportContractSerializer
+        TransportContract: TransportContractSerializer,
     }
+
 
 class GridConnectionPolymorphicSerializer(PolymorphicSerializer):
     model_serializer_mapping = {
@@ -142,8 +128,9 @@ class GridConnectionPolymorphicSerializer(PolymorphicSerializer):
         BuildingGridConnection: BuildingGridConnectionSerializer,
         ProductionGridConnection: ProductionGridConnectionSerializer,
         IndustryGridConnection: IndustryGridConnectionSerializer,
-        DistrictHeatingGridConnection: DistrictHeatingGridConnectionSerializer
+        DistrictHeatingGridConnection: DistrictHeatingGridConnectionSerializer,
     }
+
 
 class EnergyAssetPolymorphicSerializer(PolymorphicSerializer):
     model_serializer_mapping = {
@@ -168,13 +155,12 @@ class EnergyAssetPolymorphicSerializer(PolymorphicSerializer):
         StorageAsset: StorageAssetSerializer,
         HeatStorageAsset: HeatStorageAssetSerializer,
         ElectricStorageAsset: ElectricStorageAssetSerializer,
-        VehicleElectricStorageAsset: VehicleElectricStorageAssetSerializer
+        VehicleElectricStorageAsset: VehicleElectricStorageAssetSerializer,
     }
+
 
 class GridNodePolymorphicSerializer(PolymorphicSerializer):
     model_serializer_mapping = {
         ElectricGridNode: ElectricGridNodeSerializer,
-        HeatGridNode: HeatGridNodeSerializer
+        HeatGridNode: HeatGridNodeSerializer,
     }
-
-
